api/requirements.txt
 argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.9.3
attrs==21.4.0
bcrypt==3.2.0
black==22.1.0
cryptography==36.0.1
dataclasses==0.6
deprecated==1.2.13
dnspython==2.2.0
email-validator==1.1.3
flask==2.0.2
flask-jwt-extended==4.3.1
flask-login==0.5.0
flit-core==3.6.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
jinja2==3.0.3
lazy-object-proxy==1.7.1
markupsafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
mypy_extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
pip==21.2.4
platformdirs==2.3.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pyjwt==2.3.0
pylint==2.12.2
pyparsing==3.0.7
pytest==6.2.5
python-dotenv==0.19.2
redis==4.1.2
setuptools==60.5.0
six==1.16.0
toml==0.10.2
tomli==2.0.0
validate_email==1.3